TRC in go (#1309)
This PR adds TRC support to go.
Additionally,  it syncs the TRC format with the book. (See issue #933):
 - add "ExpirationTime"
 - rename "ISDID" -> "ISD"
 - rename "QuorumEEPKI" -> "ThresholdEEPKI"
 - change CertLog entry from public key to cert
 - add dict to RootCAs entries with ("Certificate", "OnlineKeyAlg", "OnlineKey", "TRCSrv", "ARPKISrv","ARPKIKey")
 - rename "RootRainsKey" -> "RAINS"
 - add dict to "RAINS" with ("RootRAINSKey", "OnlineKeyAlg", "OnlineKey", "TRCSrv")

The python code is adopted accordingly.

The code is also being prepared for TRC updates and TRC cross signatures.
Depending on the ISD of the trusted TRC, another verification method has to be used.
This is abstracted away by the verify function, which selects the right method.

For now, the generator creates empty dictionaries for the fields [RootCAs, CertLogs, RAINS],
however, parsing is supported by both python and go code.

Verification based on cross signature and verification of signatures of root CAs and Rains are not implemented yet.