Check for unique ASes in generator, use non-production ASes. (#1478)
Now that we are using global AS numbers, the generator needs to catch a
`.topo` file that repeats AS numbers. Also, we need to change to using
example AS numbers where possible, and private-use AS numbers for the
default topo etc.

For private-use, i've gone with the following scheme for our test topos:
- The prefix is 4_295_00X for ISD X.
- Core ASes have a suffix of _0X0, where X is 1,2,3 for the
  first/second/third core AS in an ISD etc.
- Non-core ASes have the suffix of _0XY where X is their core _issuer_
  AS, and Y is incremented for each non-core AS under that issuer.

topology/Default.topo is annotated with the old AS IDs to help with the
transition.

N.B.: doc/fig/default-topo.pdf needs to be updated, i'll be looking for
a volunteer. :)

For the interfaces in go/lib/xtest/graph, i've gone with the last 4
digits of each AS number, which thankfully comfortably fits in the
64bit namespace.

Ref https://github.com/scionproto/scion/wiki/ISD-and-AS-numbering

Also:
- Remove an old ignore.goconvey
- Fix address regex in snet to cope with _'s in AS numbers.