scion-pki: Make the output directory separately configurable (#1589)
Users of scion-pki can now specify an output directory for keys, certs, and TRCs using the -o/--out cmdline parameter (so far this was the same as the root directory for the configuration files). If not specified, the output directory will be the same as the root directory (-d/--root).

Fixes #1575.

Also:
Change trust_test.go to use the new capabilities.