Patch for ARM 32bit. (#1687)
Fixes some issues where 64-bit assumptions in the code cause problems on a 32-bit platform.

N.B.: There is a problem with the alignment of the sciond.connector struct in Go on ARM 32 that this does _not_ fix. https://github.com/scionproto/scion/issues/1688 is filed to track a higher-level fix for this issue.