border: unblock reader when closing a socket (#2088)
When closing a socket, ie. interface change during config reload, with
no traffic, the reader will block forever waiting for packets.

Because we apparently want to flush any in-flight packet in the border
router, although we might already lose packets received by the nic but
not deliver to the application yet, we cannot close the socket at the
start. Thus, we need SetReadDeadline funcitonality to unblock the reader
and allow the proper socket cleanup steps.

Fixes #2078