BR: Refactor socket setup (#2306)
Add `rctx.SocketType` to indicate the type of socket.
The default is `posix`.

Add `SocketConf` to configure the socket type for the
local and all interface sockets individually.

Replace hook system for local and external socket setup
with a `SocketOps` dictionary. `SocketOps` may be added
by different network stacks.
This is done in preperation for a future PR that adds
proper transactional context reloads with rollback and
teardown capabilites.

Move posix specific socket setup  to `setup-posix.go`.

Add basic unit tests for `setupNet`. These will be
extended in the future PR that adds rollback and
teardown.