BR: Fetch static topology  (#2435)
This PR enables static topology fetching from the discovery service.

The BR can be configured to periodically fetch the static topology from the discovery service and, if desired, write the updated version to disk.
Whether modifications to the semi-mutable part (i.e. the interfaces) of the topology are possible, can be enabled and disabled.

Additionally, acceptance tests are added to check that the BR fetches steady topologies as expected.