Docker base image: Use chown only when needed (#2457)
Using chown when building a docker image is expensive. It in fact copies
each file that's touched growing the size of the layer a lot.

This particular step resulted in a 84MB layer. After the change it's 422
bytes.

This patch also adds missing tzdata dependency and removes obsolete
libpcap-dev dependency.