BS: Add keepalive handler (#2588)
This PR adds the keepalive handler.

Upon receiving a IFID keepalive, the state of the interface is changed
to active. If the interface was not active before, the handler
pushes an ifid state update to all border routers to remove revocations,
starts beaconing on the interface immediately and deletes possible
revocations associated with the interfaces link (i.e. the local and
remote revocations).

fixes #2562