BS: Filter ISD loops (#2662)
Add ISD loop filtering to the policy.
By default, ISD loops are not allowed. If desired, they can be allowed
in the policy.

Additionally, add pre-check to beacon store to indicate whether a beacon
will be ingored, so that the caller does not have to spend resources on
further processing the beacon.