Gen: Leaf certificate issuing time workaround (#2801)
The generator occasionally issues leaf certificates that are valid
before the issuing Issuer certificate. This can happen, since the
granularity is one second, and the issuing time is simply current time
without considering the enclosing Issuer validity period.

This PR is a "hacky" attempt of solving the problem. As it will not take
any machine that is capable of running the topology more than a second
to generate the certificates, we simply add 2 seconds to the current
time for leaf certificates. This guarantees that the leaf certificates
validity period is covered by the Issuer certificate (given our
reasonable assumption)

To cleanly fix the issue, we should consider addressing #1491, but
this is good enough for now.