Simplify Topology-related packages and data types (#3440)
This is a non-breaking change. The topology JSON format is not
affected.

In more detail:
- deleted topopruner because it was never used; we also no longer
support non-underlay binds and sibra, so the pruning is not needed;
- moved topology types from package itopo to package topology;
- moved most JSON-specific topology code to its own package;
- added test strictly for JSON serialization/deserialization;
- added strict name tags for all topology JSON fields; this prevents
refactors from inadvertently breaking backwards compatibility
re: the topology file format;
- removed duplicated fields from topology.RWTopology (used to be
topology.Topo);
- removed some duplicated APIs from topology.Topology (used to be
itopo.Topology);
- removed TimestampHuman topology field from high-level types, the
field was never used by applications; the field is still present
in the JSON low-level type, so applications that process the
JSON contents directly can still use it;
- added documentation for all exported entities in package topology;
- renamed non-idiomatic topology type fields;
- decoupled capnp proto-specific LinkType from most applications;
- renamed overlay to underlay in most application facing APIs;
this keeps names closer to established terminology, where
SCION is the overlay and the existing IP fabric is the underlay;

Fixes https://github.com/scionproto/scion/issues/3346.