Add primary attributes to topology file (#3607)
Add the primary attributes to toplogy files. The primary attributes are described in:
https://github.com/scionproto/scion/blob/master/doc/ControlPlanePKI.md#primary-ases

This fixes the toplogy generator to handle this.

This is a breaking change, old topology file will have to be updated and replace the old `Core` boolean with the attributes array.
An easy migration strategy is to just put all attributes for previous cores, and no attributes for non-cores.